move transport into separate file and abstract away OrasRemote from progressbar
